# srcs/api-gateway/static/openapi.yaml
openapi: 3.0.0
info:
  title: Movie Platform Gateway API
  description: API Gateway interface for Inventory management and Billing processing.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Vagrant Gateway

paths:
  /api/movies:
    get:
      summary: Retrieve all movies
      description: Fetches the list of movies from the Inventory Service.
      responses:
        '200':
          description: A list of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Create a new movie
      description: Adds a new movie to the Inventory database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Movie created successfully
    delete:
      summary: Delete all movies
      responses:
        '200':
          description: Database cleared

  /api/billing:
    post:
      summary: Submit a payment order
      description: Queues a payment order to RabbitMQ for asynchronous processing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order queued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
                  message:
                    type: string
                    example: Order queued for processing

components:
  schemas:
    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
    MovieInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
    Order:
      type: object
      required:
        - user_id
        - total_amount
      properties:
        user_id:
          type: string
        number_of_items:
          type: string
        total_amount:
          type: string